Add syslog-ng related testing
Modify the process test to check for syslog-ng process instead of the
rsyslog process, because the default syslog daemon in Sulka will be
syslog-ng going forwards. Add a new filesystem test file that checks the
existence (or non-existence) of files in the root file system. This test
checks that the syslog-ng is present, and that awk-dependant programs
are removed.