Add build tests for GPLv3 code
Add build test for checking if there are GPLv3 licensed packages in the
main distro configuration. There should be none, and the build should
succeed.

Add another build test that checks that the build should fail if
FIRST_BOOT_RELABEL SELinux feature is enabled. The feature relies on
GPLv3 licensed code, so the it should not be allowed.