Add global variable file
Instead of supplying the TEMP_DIR with an argument file, use a global
variable file. This makes more sense as the TEMP_DIR is assumed to be
constant in different tests. In addition, add different kas
configuration fragment combinations as global variables to make the
tests more readable.

This also allows adding a global build configuration option
functionality to the Add Common Build Configuration keyword. This means
in practice that all the tests should have this keyword in their Suite
Setup. This allows testing different build variants, like a version of
Sulka that is using systemd.

Whe adding the global build configuration flag to the global build
variable file, remeber to escape the quotes like this:

GLOBAL_BUILD_CONFIG = "INIT_MANAGER=\"systemd\""